# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/models": {
            "get": {
                "description": "Get a list of all AI models with filtering options",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "List all models",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by provider ID",
                        "name": "provider",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by model type (chat, completion, embedding, image)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by model category",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (active, inactive, deprecated)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of models",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new AI model configuration with detailed parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "Create a new model",
                "parameters": [
                    {
                        "description": "Model configuration",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.Model"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created model",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/models/{id}": {
            "get": {
                "description": "Get detailed information about a specific model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "Get model by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Model details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Model not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing model configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "Update model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated model configuration",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.Model"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated model",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a model by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "Delete model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/chat/completions": {
            "post": {
                "description": "Create a chat completion using OpenAI-compatible API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openai"
                ],
                "summary": "Create chat completion",
                "parameters": [
                    {
                        "description": "Chat completion request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ChatCompletionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat completion response",
                        "schema": {
                            "$ref": "#/definitions/handler.ChatCompletionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/completions": {
            "post": {
                "description": "Create a text completion using OpenAI-compatible API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openai"
                ],
                "summary": "Create completion",
                "parameters": [
                    {
                        "description": "Completion request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CompletionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Completion response",
                        "schema": {
                            "$ref": "#/definitions/handler.CompletionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/models/{model}": {
            "get": {
                "description": "Get information about a specific model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openai"
                ],
                "summary": "Get model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model ID",
                        "name": "model",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Model information",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ChatCompletionRequest": {
            "description": "OpenAI聊天补全请求结构",
            "type": "object",
            "required": [
                "messages",
                "model"
            ],
            "properties": {
                "frequency_penalty": {
                    "type": "number"
                },
                "max_tokens": {
                    "type": "integer"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.ChatMessage"
                    }
                },
                "model": {
                    "type": "string"
                },
                "presence_penalty": {
                    "type": "number"
                },
                "stop": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stream": {
                    "type": "boolean"
                },
                "temperature": {
                    "type": "number"
                },
                "top_p": {
                    "type": "number"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "handler.ChatCompletionResponse": {
            "description": "OpenAI聊天补全响应结构",
            "type": "object",
            "properties": {
                "choices": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "finish_reason": {
                                "type": "string"
                            },
                            "index": {
                                "type": "integer"
                            },
                            "message": {
                                "$ref": "#/definitions/handler.ChatMessage"
                            }
                        }
                    }
                },
                "created": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "usage": {
                    "type": "object",
                    "properties": {
                        "completion_tokens": {
                            "type": "integer"
                        },
                        "prompt_tokens": {
                            "type": "integer"
                        },
                        "total_tokens": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
...
        "handler.CompletionRequest": {
            "description": "OpenAI文本补全请求结构",
            "type": "object",
            "required": [
                "model",
                "prompt"
            ],
            "properties": {
                "frequency_penalty": {
                    "type": "number"
                },
                "max_tokens": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "presence_penalty": {
                    "type": "number"
                },
                "prompt": {
                    "type": "string"
                },
                "stop": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stream": {
                    "type": "boolean"
                },
                "temperature": {
                    "type": "number"
                },
                "top_p": {
                    "type": "number"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "handler.CompletionResponse": {
            "type": "object",
            "properties": {
                "choices": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "finish_reason": {
                                "type": "string"
                            },
                            "index": {
                                "type": "integer"
                            },
                            "text": {
                                "type": "string"
                            }
                        }
                    }
                },
                "created": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "usage": {
                    "type": "object",
                    "properties": {
                        "completion_tokens": {
                            "type": "integer"
                        },
                        "prompt_tokens": {
                            "type": "integer"
                        },
                        "total_tokens": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
...
        "storage.Model": {
            "type": "object",
            "properties": {
                "capabilities": {
                    "description": "模型能力: text-generation, code-completion, image-analysis, etc.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "category": {
                    "description": "模型分类: gpt, claude, gemini, llama, etc.",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "features": {
                    "description": "功能特性",
                    "allOf": [
                        {
                            "$ref": "#/definitions/storage.ModelFeatures"
                        }
                    ]
                },
                "id": {
                    "type": "string"
                },
                "limits": {
                    "description": "使用限制",
                    "allOf": [
                        {
                            "$ref": "#/definitions/storage.ModelLimits"
                        }
                    ]
                },
                "metadata": {
                    "description": "扩展元数据",
                    "type": "object",
                    "additionalProperties": true
                },
                "model_id": {
                    "description": "提供商内部的模型标识",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parameters": {
                    "description": "默认参数",
                    "allOf": [
                        {
                            "$ref": "#/definitions/storage.ModelParameters"
                        }
                    ]
                },
                "pricing": {
                    "description": "定价信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/storage.ModelPricing"
                        }
                    ]
                },
                "provider_id": {
                    "type": "string"
                },
                "status": {
                    "description": "active, inactive, deprecated",
                    "type": "string"
                },
                "type": {
                    "description": "模型类型: chat, completion, embedding, image",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "version": {
                    "description": "模型版本",
                    "type": "string"
                }
            }
        },
        "storage.ModelFeatures": {
            "type": "object",
            "properties": {
                "code_completion": {
                    "description": "是否支持代码补全",
                    "type": "boolean"
                },
                "embeddings": {
                    "description": "是否支持嵌入",
                    "type": "boolean"
                },
                "fine_tuning": {
                    "description": "是否支持微调",
                    "type": "boolean"
                },
                "function_calling": {
                    "description": "是否支持函数调用",
                    "type": "boolean"
                },
                "moderation": {
                    "description": "是否支持内容审核",
                    "type": "boolean"
                },
                "multilingual": {
                    "description": "是否支持多语言",
                    "type": "boolean"
                },
                "streaming": {
                    "description": "是否支持流式响应",
                    "type": "boolean"
                },
                "vision": {
                    "description": "是否支持视觉理解",
                    "type": "boolean"
                }
            }
        },
        "storage.ModelLimits": {
            "type": "object",
            "properties": {
                "context_window": {
                    "description": "上下文窗口大小",
                    "type": "integer"
                },
                "max_frequency_penalty": {
                    "type": "number"
                },
                "max_input_tokens": {
                    "description": "最大输入token数",
                    "type": "integer"
                },
                "max_output_tokens": {
                    "description": "最大输出token数",
                    "type": "integer"
                },
                "max_presence_penalty": {
                    "type": "number"
                },
                "max_temperature": {
                    "description": "最大温度值",
                    "type": "number"
                },
                "max_tokens": {
                    "description": "最大token数",
                    "type": "integer"
                },
                "max_top_p": {
                    "description": "最大top-p值",
                    "type": "number"
                },
                "min_temperature": {
                    "description": "最小温度值",
                    "type": "number"
                }
            }
        },
...
    }
}